# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is product api swagger doc.",
        "title": "Swagger Product API",
        "contact": {
            "name": "yan.jinbin",
            "email": "yan.jinbin@shipper.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.0.1"
    },
    "host": "127.0.0.1:10000",
    "basePath": "/v1",
    "paths": {
        "/v1/brands": {
            "get": {
                "description": "Marketplace open api doc link. Lazada: https://bit.ly/3l8loTF shopee: https://bit.ly/3L7JMQ5 Tokopedia: Tiktok: https://bit.ly/3FO6Pym\ne.g. get brand list like Apple Google Facebook Airbnb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brand"
                ],
                "summary": "get brand list from marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization signature",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mock-positive",
                            "mock-negative"
                        ],
                        "type": "string",
                        "description": "test-id Key via Header",
                        "name": "x-test-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "common query params: shop id",
                        "name": "shop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: marketplace type",
                        "name": "mp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: page number",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTBrandListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/v1/categories": {
            "get": {
                "description": "Marketplace open api doc link. Lazada: https://bit.ly/3wbFq64, shopee: https://bit.ly/3l7cjKT, Tokopedia: https://bit.ly/39fr2Rj, Tiktok: https://bit.ly/3FI0HaM\ne.g. get category tree, like telecommunication-\u003emobile phone-\u003e iphone -\u003e iphone 13 series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get category tree or dictionary from marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization signature",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mock-positive",
                            "mock-negative"
                        ],
                        "type": "string",
                        "description": "test-id Key via Header",
                        "name": "x-test-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "common query params: shop id",
                        "name": "shop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: marketplace type",
                        "name": "mp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPCategoryTreeResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/v1/categories/{category_id}/attributes": {
            "get": {
                "description": "Marketplace open api doc link. Lazada: https://bit.ly/3Ncd0yp  shopee: https://bit.ly/3PxuYOb  Tokopedia: https://bit.ly/3Lwltvj Tiktok: https://bit.ly/3wtQdrm\ne.g. under category_id 100, product attributes, include mandatory or not, are waterproof fireproof isDangerous R18",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "get category attributes from marketplace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization signature",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mock-positive",
                            "mock-negative"
                        ],
                        "type": "string",
                        "description": "test-id Key via Header",
                        "name": "x-test-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "common query params: shop id",
                        "name": "shop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: marketplace type",
                        "name": "mp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPProductAttrsResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    }
                }
            }
        },
...
        "/v1/products/active": {
            "post": {
                "description": "Marketplace open api doc link. Lazada: https://bit.ly/37Ix6kR shopee: https://bit.ly/3wbm6G6 Tokopedia: https://bit.ly/3wkDsiH Tiktok: https://bit.ly/3sxkOTu\ne.g. iphone 13 red 256GB is top seller goods. so we often activate it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "batch activate products on shelf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization signature",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mock-positive",
                            "mock-negative"
                        ],
                        "type": "string",
                        "description": "test-id Key via Header",
                        "name": "x-test-id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "common query params: shop id",
                        "name": "shop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: marketplace type",
                        "name": "mp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "activate Product From Shelf",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ProductIDSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPEmptyResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/v1/products/inactive": {
            "post": {
                "description": "Marketplace open api doc link. Lazada: https://bit.ly/37Ix6kR shopee: https://bit.ly/3wbm6G6 Tokopedia: https://bit.ly/3wkDsiH Tiktok: https://bit.ly/3sxkOTu\ne.g. iphone 13 red 256GB is top seller goods. so we often deactivate it for the sake of shortage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "batch deactivate products from shelf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization signature",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mock-positive",
                            "mock-negative"
                        ],
                        "type": "string",
                        "description": "test-id Key via Header",
                        "name": "x-test-id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "common query params: shop id",
                        "name": "shop_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "common query params: marketplace type",
                        "name": "mp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "deactivate Product From Shelf",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ProductIDSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPEmptyResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.HTTPErrResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.BrandListData": {
            "type": "object",
            "properties": {
                "brand_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tiktok.Brand"
                    }
                }
            }
        },
        "entity.HTTBrandListResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.BrandListData"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Meta"
                },
                "pagination": {
                    "$ref": "#/definitions/entity.Pagination"
                }
            }
        },
...
        "entity.HTTPCategoryTreeResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tiktok.Category"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Meta"
                }
            }
        },
        "entity.HTTPEmptyResp": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/entity.Meta"
                }
            }
        },
        "entity.HTTPErrResp": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/entity.Meta"
                }
            }
        },
        "entity.HTTPProductAttrsResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.ProductAttrsData"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Meta"
                }
            }
        },
...
        "entity.Meta": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "entity.ProductIDSRequest": {
            "type": "object",
            "properties": {
                "product_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "tiktok.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_leaf": {
                    "type": "boolean"
                },
                "local_display_name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
...
    }
}